refactor to move into utils and make tests more robust
