+++ tests/integration_tests/test_hyperopt_ray_horovod.py
@@ -22,15 +22,11 @@ from packaging import version
 
 from ludwig.api import LudwigModel
 from ludwig.callbacks import Callback
-from ludwig.constants import ACCURACY, EXECUTOR, NUM_SAMPLES, PARAMETERS, TRAINER
+from ludwig.constants import ACCURACY, EXECUTOR, TRAINER
 from ludwig.globals import HYPEROPT_STATISTICS_FILE_NAME
 from ludwig.hyperopt.results import HyperoptResults
-from ludwig.hyperopt.run import (
-    get_total_trial_count,
-    hyperopt,
-    update_hyperopt_params_with_defaults,
-    update_or_set_max_concurrent_trials,
-)
+from ludwig.hyperopt.run import hyperopt, update_hyperopt_params_with_defaults
+from ludwig.hyperopt.utils import update_or_set_max_concurrent_trials
 from ludwig.utils.defaults import merge_with_defaults
 from tests.integration_tests.utils import binary_feature, create_data_set_to_use, generate_data, number_feature
 
@@ -188,11 +184,7 @@ def run_hyperopt_executor(
 
     backend = RayBackend(**RAY_BACKEND_KWARGS)
     update_hyperopt_params_with_defaults(hyperopt_config)
-    update_or_set_max_concurrent_trials(
-        hyperopt_config[EXECUTOR],
-        backend,
-        get_total_trial_count(hyperopt_config[PARAMETERS], hyperopt_config[EXECUTOR].get(NUM_SAMPLES)),
-    )
+    update_or_set_max_concurrent_trials(hyperopt_config, backend)
 
     parameters = hyperopt_config["parameters"]
     if search_alg.get("type", "") == "bohb":
