+++ tests/ludwig/hyperopt/test_hyperopt.py
@@ -2,8 +2,7 @@ import pytest
 
 from ludwig.backend import initialize_backend
 from ludwig.constants import INPUT_FEATURES, NAME, OUTPUT_FEATURES, TYPE
-from ludwig.hyperopt.run import update_or_set_max_concurrent_trials
-from ludwig.hyperopt.utils import substitute_parameters
+from ludwig.hyperopt.utils import substitute_parameters, update_or_set_max_concurrent_trials
 
 BASE_CONFIG = {
     INPUT_FEATURES: [{NAME: "title", TYPE: "text"}],
@@ -79,24 +78,64 @@ def test_substitute_parameters(parameters, expected):
     assert actual_config == expected
 
 
+# @pytest.mark.parametrize("parameters, expected", [], ids=[])
+# def test_get_total_trial_count(parameters, expected):
+#     pass
+
+
 @pytest.mark.distributed
 @pytest.mark.parametrize(
     "parameters, expected",
     [
-        (
-            {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": None},
-            {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": None},
+        (  # If max_concurrent_trials is none, it should not be set in the updated config
+            {
+                "parameters": {"trainer.learning_rate": {"space": "choice", "values": [0.001, 0.01, 0.1]}},
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": None},
+            },
+            {
+                "parameters": {"trainer.learning_rate": {"space": "choice", "values": [0.001, 0.01, 0.1]}},
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": None},
+            },
         ),
-        (
-            {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": "auto"},
-            {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": 2},
+        (  # If max_concurrent_trials is auto, set it to total_trials - 2 if num_samples == num_cpus
+            {
+                "parameters": {"trainer.learning_rate": {"space": "choice", "values": [0.001, 0.01, 0.1]}},
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": "auto"},
+            },
+            {
+                "parameters": {"trainer.learning_rate": {"space": "choice", "values": [0.001, 0.01, 0.1]}},
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": 2},
+            },
         ),
-        (
-            {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": 1},
-            {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": 1},
+        (  # Even though num_samples is set to 4, this will actually result in 9 trials. This test checks if
+            # we still correctly set max_concurrent_trials to 2
+            {
+                "parameters": {
+                    "trainer.learning_rate": {"space": "grid_search", "values": [0.001, 0.01, 0.1]},
+                    "combiner.num_fc_layers": {"space": "grid_search", "values": [1, 2, 3]},
+                },
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": "auto"},
+            },
+            {
+                "parameters": {
+                    "trainer.learning_rate": {"space": "grid_search", "values": [0.001, 0.01, 0.1]},
+                    "combiner.num_fc_layers": {"space": "grid_search", "values": [1, 2, 3]},
+                },
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": 2},
+            },
+        ),
+        (  # Ensure user config value (1) is respected if it is less than total possible trials (2)
+            {
+                "parameters": {"trainer.learning_rate": {"space": "choice", "values": [0.001, 0.01, 0.1]}},
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": 1},
+            },
+            {
+                "parameters": {"trainer.learning_rate": {"space": "choice", "values": [0.001, 0.01, 0.1]}},
+                "executor": {"num_samples": 4, "cpu_resources_per_trial": 1, "max_concurrent_trials": 1},
+            },
         ),
     ],
-    ids=["none", "auto", "1"],
+    ids=["none", "auto", "auto_with_large_num_trials", "1"],
 )
 def test_set_max_concurrent_trials(parameters, expected, ray_cluster_4cpu):
     backend = initialize_backend("ray")
